+++ src/main/java/com/github/toolarium/jwebserver/util/ConfigurationUtil.java
@@ -5,12 +5,23 @@
  */
 package com.github.toolarium.jwebserver.util;
 
+import com.github.toolarium.common.security.ISecuredValue;
+import com.github.toolarium.common.security.SecuredValue;
+import com.github.toolarium.common.util.PropertyExpander;
+import com.github.toolarium.jwebserver.logger.VerboseLevel;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+
 /**
  * The configuration util
  * 
  * @author patrick
  */
 public final class ConfigurationUtil {
+    private static final Logger LOG = LoggerFactory.getLogger(ConfigurationUtil.class);
+    private static final String END_VALUE = "].";
+
 
     /**
      * Private class, the only instance of the singelton which will be created by accessing the holder class.
@@ -40,6 +51,134 @@ public final class ConfigurationUtil {
     }
 
     
+    /**
+     * Expand a string and replace environment variables or system properties
+     *
+     * @param value the value
+     * @return the expanded value
+     */
+    public String expand(String value) {
+        if (value == null) {
+            return value;
+        }
+        
+        final String result = PropertyExpander.getInstance().expand(value);
+        if (!value.equals(result)) {
+            LOG.debug("Resolved value [" + value + "] to [" + result + "].  ");
+        }
+        
+        return result;
+    }
+
+
+    /**
+     * Expand a string and replace environment variables or system properties
+     *
+     * @param valueList the value list
+     * @return the expanded value list
+     */
+    public String[] expand(String[] valueList) {
+        if (valueList == null || valueList.length == 0) {
+            return valueList;
+        }
+        
+        String[] expandedList = new String[valueList.length];
+        for (int i = 0; i < expandedList.length; i++) {
+            expandedList[i] = expand(valueList[i]);
+        }
+
+        return expandedList;
+    }
+    
+
+    /**
+     * Convert integer
+     *
+     * @param name the attribute name
+     * @param value the value
+     * @param defaultValue the default value
+     * @return the value
+     */
+    public Integer convert(String name, String value, Integer defaultValue) {
+        try {
+            return Integer.valueOf(expand(expand(value)));
+        } catch (Exception e) {
+            String nameMsg = ""; 
+            if (name != null) {
+                nameMsg = "for attribute [" + name + "]";
+            }
+            LOG.warn("Invalid value [" + value + "] " + nameMsg + ", keep default value [" + defaultValue + END_VALUE);
+            return defaultValue;
+        }
+    }
+
+    
+    /**
+     * Convert boolean
+     *
+     * @param name the attribute name
+     * @param value the value
+     * @param defaultValue the default value
+     * @return the value
+     */
+    public Boolean convert(String name, String value, Boolean defaultValue) {
+        try {
+            return Boolean.valueOf(expand(value));
+        } catch (Exception e) {
+            String nameMsg = ""; 
+            if (name != null) {
+                nameMsg = "for attribute [" + name + "]";
+            }
+            LOG.warn("Invalid value [" + value + "] " + nameMsg + ", keep default value [" + defaultValue + END_VALUE);
+            return defaultValue;
+        }
+    }
+
+
+    /**
+     * Convert verbose level
+     *
+     * @param name the attribute name
+     * @param value the value
+     * @param defaultValue the default value
+     * @return the value
+     */
+    public VerboseLevel convert(String name, String value, VerboseLevel defaultValue) {
+        try {
+            return VerboseLevel.valueOf(expand(value));
+        } catch (Exception e) {
+            String nameMsg = ""; 
+            if (name != null) {
+                nameMsg = "for attribute [" + name + "]";
+            }
+            LOG.warn("Invalid value [" + value + "] " + nameMsg + ", keep default value [" + defaultValue + END_VALUE);
+            return defaultValue;
+        }
+    }
+
+    
+    /**
+     * Convert secured value
+     *
+     * @param name the attribute name
+     * @param value the value
+     * @param defaultValue the default value
+     * @return the value
+     */
+    public ISecuredValue<String> convert(String name, String value, ISecuredValue<String> defaultValue) {
+        try {
+            return new SecuredValue<String>(expand(value));
+        } catch (Exception e) {
+            String nameMsg = ""; 
+            if (name != null) {
+                nameMsg = "for attribute [" + name + "]";
+            }
+            LOG.warn("Invalid value [" + value + "] " + nameMsg + ", keep default value [" + defaultValue + END_VALUE);
+            return defaultValue;
+        }
+    }
+
+
     /**
      * Parse the string array
      *