+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [ 1.2.1 ] - 2024-07-28
+### Added
+- Added support to expand/resolve proxy hostname list by environment variables or java system properties.
 
 ## [ 1.2.0 ] - 2024-07-28
 ### Added